dbus.inc: Move dbus-daemon from auto package utils to dbus
The dbus recipe split out dbus-daemon with auto package utils, which
resultet in a "dbus" package with scripts, configs, and folder layout;
and a "util/dbus-daemon" with the daemon binary itself. Neither make
much sense without the other, so remove dbus-daemon from auto package
utils, so that it is included in the dbus package itself.

The removal led to missing build time dependencies, which are added to
DEPEND_${PA}.

The dbus init script relies on dbus-uuidgen, so add this as a dependency
to the dbus package.

Signed-off-by: Martin Hundebøll <mnhu@prevas.dk>