c-ares: new recipe
Add a recipe for the c-ares library used for async dns lookups. This is
needed to build nodejs, and thus we add a nodejs patch to make it
compatible.

The patch to make nodejs happy about record_start member of tx-structs
is required for all targets.

Explicitly set CFLAGS to avoid failing configure due to U_FORTIFY_SOURCE
not being specified in CPPFLAGS only.